Herencia y Polimorfismo - Implementación del Método calcularSalario() en Cla Vendedor por Miguel Sánchez

+++ src/Gerente.java
@@ -1,5 +1,5 @@
 public class Gerente extends Empleado{
-    public Gerente(String cargo, String nombre, Integer salario) {
+    public Gerente(String cargo, String nombre, Float salario) {
         super(cargo, nombre, salario);
     }
 }
